Chunk vertices are packed to 32bit integers

+++ KuchCraft/assets/shaders/chunk.frag.glsl
@@ -6,9 +6,9 @@ uniform sampler2D u_Textures[##max_texture_slots];
 ##world_data_uniform_buffer
 
 in vec2 v_TexCoord;
-in flat float v_TexIndex;
+in flat uint v_TexIndex;
 
 void main()
 {
-	color = texture(u_Textures[int(v_TexIndex)], v_TexCoord) * u_TintColor;
+	color = texture(u_Textures[v_TexIndex], v_TexCoord) * u_TintColor;
 }
